refactor validWaypoint
